add retry when batch fails
